Move DocumentUrlService functionality to a plugin
