+++ lms/product/canvas/__init__.py
@@ -1,5 +1,6 @@
 from lms.product.canvas._plugin.course_copy import CanvasCourseCopyPlugin
 from lms.product.canvas._plugin.grouping import CanvasGroupingPlugin
+from lms.product.canvas._plugin.misc import CanvasMiscPlugin
 from lms.product.canvas.product import Canvas
 
 
@@ -12,3 +13,4 @@ def includeme(config):  # pragma: nocover
     config.register_service_factory(
         CanvasCourseCopyPlugin.factory, iface=CanvasCourseCopyPlugin
     )
+    config.register_service_factory(CanvasMiscPlugin.factory, iface=CanvasMiscPlugin)
