+++ lms/product/canvas/product.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 
 from lms.product.canvas._plugin.course_copy import CanvasCourseCopyPlugin
 from lms.product.canvas._plugin.grouping import CanvasGroupingPlugin
+from lms.product.canvas._plugin.misc import CanvasMiscPlugin
 from lms.product.product import PluginConfig, Product, Routes
 
 
@@ -17,7 +18,9 @@ class Canvas(Product):
     )
 
     plugin_config: PluginConfig = PluginConfig(
-        grouping=CanvasGroupingPlugin, course_copy=CanvasCourseCopyPlugin
+        grouping=CanvasGroupingPlugin,
+        course_copy=CanvasCourseCopyPlugin,
+        misc=CanvasMiscPlugin,
     )
 
     settings_key = "canvas"
