+++ tests/unit/lms/views/lti/basic_launch_test.py
@@ -14,12 +14,12 @@ from tests import factories
 
 class TestHasDocumentURL:
     @pytest.mark.parametrize("document_url", (None, "a_url"))
-    def test_it(self, document_url_service, pyramid_request, document_url):
-        document_url_service.get_document_url.return_value = document_url
+    def test_it(self, misc_plugin, pyramid_request, document_url):
+        misc_plugin.get_document_url.return_value = document_url
 
         result = has_document_url(sentinel.context, pyramid_request)
 
-        document_url_service.get_document_url.assert_called_once_with(pyramid_request)
+        misc_plugin.get_document_url.assert_called_once_with(pyramid_request)
         assert result == bool(document_url)
 
 
@@ -174,17 +174,17 @@ class TestBasicLaunchViews:
     def test_configured_launch(
         self,
         svc,
-        document_url_service,
+        misc_plugin,
         pyramid_request,
         _show_document,
         LTIEvent,
     ):
         svc.configured_launch()
 
-        document_url_service.get_document_url.assert_called_once_with(pyramid_request)
+        misc_plugin.get_document_url.assert_called_once_with(pyramid_request)
 
         _show_document.assert_called_once_with(
-            document_url=document_url_service.get_document_url.return_value
+            document_url=misc_plugin.get_document_url.return_value
         )
         LTIEvent.assert_called_once_with(
             request=pyramid_request,
